Extract SendUpdates from Contribute sidebar
