+++ app/i18n/en/topics.js
@@ -47,12 +47,14 @@ const topics = {
     },
     contribute: {
       header: 'Contribute',
-      info: 'You can send updates back to <0>{{upstreamTopicTitle}}</0>. The topic owner and collaborators will be able to see the updates you\'ve sent, and after reviewing, these changes can be integrated in <0>{{upstreamTopicTitle}}</0>. You can check back here on the status of your contribution.',
-      count: 'This contribution will include {{count}} updates.',
-      saveChanges: 'Please save your changes first',
+      sendUpdates: {
+        info: 'You can send updates back to <0>{{upstreamTopicTitle}}</0>. The topic owner and collaborators will be able to see the updates you\'ve sent, and after reviewing, these changes can be integrated in <0>{{upstreamTopicTitle}}</0>. You can check back here on the status of your contribution.',
+        count: 'This contribution will include {{count}} updates.',
+        saveChanges: 'Please save your changes first',
+      },
       pending: {
         title: 'Pending contributions',
-        empty: 'You have not sent any contributions',
+        empty: 'You have not sent any updates',
       },
     },
   },
